+++ cinetica/__init__.py
@@ -1,6 +1,6 @@
 from . import cinematica
 
-__version__ = "0.10.3"
+__version__ = "0.10.4"
 
 __all__ = [
     "cinematica",