+++ cinetica/cinematica/__init__.py
@@ -1,17 +1,17 @@
-from .circular.mcu import MovimientoCircularUniforme
-from .circular.mcuv import MovimientoCircularUniformementeVariado
-
-# Import other modules that should be part of cinetica.cinematica
-from .oscilatorio.mac import MovimientoArmonicoComplejo
-from .oscilatorio.mas import MovimientoArmonicoSimple
+from .base_movimiento import Movimiento
+from .circular.movimiento_circular_uniforme import MovimientoCircularUniforme
+from .circular.movimiento_circular_uniformemente_variado import MovimientoCircularUniformementeVariado
+from .oscilatorio.movimiento_armonico_complejo import MovimientoArmonicoComplejo
+from .oscilatorio.movimiento_armonico_simple import MovimientoArmonicoSimple
 from .parabolico.analisis import MovimientoParabolicoAnalisis
 from .parabolico.base import MovimientoParabolicoBase
-from .rectilineo.mru import MovimientoRectilineoUniforme
-from .rectilineo.mruv import MovimientoRectilineoUniformementeVariado
+from .rectilineo.movimiento_rectilineo_uniforme import MovimientoRectilineoUniforme
+from .rectilineo.movimiento_rectilineo_uniformemente_variado import MovimientoRectilineoUniformementeVariado
 from .relativo.velocidad_relativa import MovimientoRelativo
 from .espacial.movimiento_espacial import MovimientoEspacial
 
 __all__ = [
+    "Movimiento",
     "MovimientoCircularUniforme",
     "MovimientoCircularUniformementeVariado",
     "MovimientoArmonicoComplejo",