+++ tests/test_cinetica.py
@@ -3,4 +3,4 @@ import pytest
 from cinetica import __version__
 
 def test_version():
-    assert __version__ == "0.10.3"
+    assert __version__ == "0.10.4"