+++ usage/movimiento_rectilineo_uniformemente_variado_ejemplo.py
@@ -4,7 +4,7 @@ from cinetica.cinematica.rectilineo.movimiento_rectilineo_uniformemente_variado 
 mruv = MovimientoRectilineoUniformementeVariado(posicion_inicial=0.0, velocidad_inicial=10.0, aceleracion_inicial=2.0)
 posicion_mruv = mruv.posicion(tiempo=3.0)
 velocidad_mruv = mruv.velocidad(tiempo=3.0)
-aceleracion_mruv = mruv.aceleracion()
+aceleracion_mruv = mruv.aceleracion(tiempo=3.0)
 print(f"MRUV - Posición a los 3s: {posicion_mruv} m, Velocidad: {velocidad_mruv} m/s, Aceleración: {aceleracion_mruv} m/s^2")
 
 # MRUV sin tiempo
@@ -26,6 +26,6 @@ print(f"MRUV - Tiempo para velocidad 20m/s: {tiempo_velocidad:.2f}s")
 # Ejemplo: Calcular el desplazamiento para alcanzar una velocidad final de 20 m/s
 desplazamiento_sin_tiempo = mruv.desplazamiento_sin_tiempo(velocidad_final=20.0)
 print(f"MRUV - Desplazamiento sin tiempo (para velocidad 20m/s): {desplazamiento_sin_tiempo:.2f}m")
-mruv.graficar(t_max=10) # Ejemplo de graficación
+# mruv.graficar(t_max=10) # Graficación se manejará por una clase Graficador separada
 
 # hspace = 0.486